+++ Database/Migrations/2020_11_12_111936_add_sort_order_in_product_table.php
@@ -26,7 +26,9 @@ class AddSortOrderInProductTable extends Migration
     public function down()
     {
       Schema::table('icommerce__products', function (Blueprint $table) {
-        $table->dropColumn('sort_order');
+        if(Schema::hasColumn('icommerce__products','sort_order')) {
+          $table->dropColumn('sort_order');
+        }
       });
     }
 }
