+++ Database/Migrations/2020_11_13_004239_add_status_in_category_table.php
@@ -26,7 +26,9 @@ class AddStatusInCategoryTable extends Migration
     public function down()
     {
         Schema::table('icommerce__categories', function (Blueprint $table) {
-					$table->dropColumn('status');
+          if(Schema::hasColumn('icommerce__categories','status')) {
+            $table->dropColumn('status');
+          }
         });
     }
 }
