+++ Database/Migrations/2021_01_13_199988_add_parent_in_payment_methods_table.php
@@ -26,7 +26,9 @@ class AddParentInPaymentMethodsTable extends Migration
     public function down()
     {
         Schema::table('icommerce__payment_methods', function (Blueprint $table) {
-			$table->dropColumn('parent_name');
+          if(Schema::hasColumn('icommerce__payment_methods','parent_name')) {
+            $table->dropColumn('parent_name');
+          }
         });
     }
 }
